Makefile: Makefile: only preserve links when installing output files
When installing output files with the `install` Makefile target, the
output file ownership is preserved because the file copy to the
destination directory is done with `cp -a/--archive`.

When using the `install` Makefile target in a Yocto recipe, it triggers
the following Bitbake host contamination warning:

WARNING: optee-client-3.17.0.imx-r0 do_package_qa: QA Issue: optee-client: /usr/lib/libseteec.so is owned by uid 1001, which is the same as the user running bitbake. This may be due to host contamination
optee-client: /usr/lib/libteec.so is owned by uid 1001, which is the same as the user running bitbake. This may be due to host contamination

The original goal of `cp -a` was to preserve symbolics links to shared
libraries. The issue is that it also preserves ownership which is
unwanted. To preserve symbolic links, `cp -d` is enough.

Fixes: 44c0d3056 ("Makefile: preserve links to libraries during make copy_export")
Signed-off-by: Clement Faure <clement.faure@nxp.com>
Reviewed-by: Jerome Forissier <jerome.forissier@linaro.org>
(cherry picked from commit a5c30b1277466a9bf85b62f45a6b00e79774e29c)